todo_service/requirements.txt
fastapi~=0.92.0
pydantic~=1.10.5
uvicorn~=0.20.0
starlette~=0.25.0
python-dotenv~=1.0.0
SQLAlchemy~=2.0.23
psycopg2-binary~=2.9.1
python-multipart~=0.0.6